perf(list): better concurrency management with time-based etag
