+++ apps/client/src/app/modules/list/+state/lists.reducer.ts
@@ -2,7 +2,6 @@ import { ListsAction, ListsActionTypes } from './lists.actions';
 import { List } from '../model/list';
 import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
 import { ListController } from '../list-controller';
-import { cloneDeep } from 'lodash';
 
 
 const PINNED_LIST_LS_KEY = 'lists:pinned';
@@ -182,7 +181,7 @@ export function listsReducer(
     }
 
     case ListsActionTypes.ListDetailsLoaded: {
-      const newVersion = ListController.clone(cloneDeep(action.payload) as List, true);
+      const newVersion = ListController.clone(structuredClone(action.payload) as List, true);
       let listDetails = state.listDetails;
       if (state.listDetails.entities[action.payload.$key]) {
         listDetails = listsAdapter.mapOne({
